Added GSM

+++ Files/Header/lm35.h
@@ -1,2 +1,2 @@
 //lm35.h
-u8 ReadTempLM35(u8 chNo);
+f32 ReadTempLM35(u8 chNo);
